Computed Properties

+++ livewire-app/app/Livewire/ArticleList.php
@@ -4,6 +4,7 @@ namespace App\Livewire;
 
 use Livewire\Component;
 use Livewire\Attributes\Title;
+use Livewire\Attributes\Computed;
 use App\Models\Article;
 use Livewire\WithPagination;
 
@@ -14,9 +15,27 @@ class ArticleList extends AdminComponent
 
     public $showOnlyPublished = false;
 
+    #[Computed/*(persist: true)*/]
+    public function articles()
+    {
+        $query = Article::query();
+
+        if ($this->showOnlyPublished) {
+            $query->where('published', 1);
+        }
+
+        return $query->paginate(10, pageName: 'articles-page');
+    }
+
     public function delete(Article $article)
     {
+        if ($this->articles->count() < 10 ) {
+            throw new \Exception("Nope");
+        }
+
         $article->delete();
+        unset($this->articles);
+        cache()->forget('published-count');
     }
 
     public function showAll()
@@ -29,18 +48,5 @@ class ArticleList extends AdminComponent
     {
         $this->showOnlyPublished = true;
         $this->resetPage(pageName: 'articles-page');
-    }
-
-    public function render()
-    {
-        $query = Article::query();
-
-        if ($this->showOnlyPublished) {
-            $query->where('published', 1);
-        }
-
-        return view('livewire.article-list', [
-            'articles' => $query->paginate(10, pageName: 'articles-page'),
-        ]);
     }
 }
